Prevent multiple annotations being opened
